backend/requirements.txt
Flask
firebase-admin
requests
python-dotenv
flask-cors